requirements.txt
asgiref==3.3.1
cycler==0.10.0
Django==3.1.6
django-appconf==1.0.4
django-debug-toolbar==3.2
django-summernote==0.8.11.6
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.1
opencv-python==4.5.1.48
pilkit==2.0
Pillow==8.1.0
progressbar2==3.53.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-utils==2.5.6
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
